analytics and advanced analytics components updated
